Type EPOCH datetimes as Long in Kotlin and Scala

+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Kotlin_record_epoch_i32_overflow@v1_9.kts
@@ -1,5 +1,5 @@
-data class Record0(val within_i32: Int, val beyond_i32: Long)
+data class Record0(val within_i32: Long, val beyond_i32: Long)
 val my_data = Record0(
-    within_i32 = 1705320000,
-    beyond_i32 = 4085195400,
+    within_i32 = 1705320000L,
+    beyond_i32 = 4085195400L,
 )
